requirements.txt
Flask
pdf2image
Pillow
PyYAML
pydantic
requests
